refactor: pull out LibraryList and PlayerStore
