+++ src/renderer/store/main.ts
@@ -8,19 +8,27 @@ interface AdditionalActions {
   setLibraryPath: (path: string) => void;
 }
 
-const useStoreStructure = create<StoreStructure & AdditionalActions>((set) => ({
+const useMainStore = create<StoreStructure & AdditionalActions>((set) => ({
+  /**
+   * StoreStructure
+   */
   library: {},
   playlists: [],
   lastPlayedSong: '',
   libraryPath: '',
+
+  /**
+   * AdditionalActions
+   */
   deleteEverything: () => set({}, true),
   setLibrary: (library: { [key: string]: SongSkeletonStructure }) => {
+    // @dev: library source of truth is the BE so this only updates the FE store
     return set({
       library,
     });
   },
   setLastPlayedSong: (song: string) => {
-    // send the message over the IPC to sync the backend with the FE
+    // @dev: send message that syncs the BE with the FE, affects next reboot
     window.electron.ipcRenderer.sendMessage('set-last-played-song', song);
     return set({
       lastPlayedSong: song,
@@ -33,4 +41,4 @@ const useStoreStructure = create<StoreStructure & AdditionalActions>((set) => ({
   },
 }));
 
-export default useStoreStructure;
+export default useMainStore;
